+++ ai/MiniSimulator.cs
@@ -25,7 +25,7 @@ namespace HREngine.Bots
         public int dirtyTwoTurnSim = 256;
 
         public Action bestmove = null;
-        public int bestmoveValue = 0;
+        public float bestmoveValue = 0;
         public Playfield bestboard = new Playfield();
 
         public Behavior botBase = null;
@@ -92,7 +92,7 @@ namespace HREngine.Bots
             }
         }
 
-        public int doallmoves(Playfield playf, bool isLethalCheck)
+        public float doallmoves(Playfield playf, bool isLethalCheck)
         {
             //Helpfunctions.Instance.logg("NXTTRN" + playf.mana);
             if (botBase == null) botBase = Ai.Instance.botBase;
@@ -113,7 +113,7 @@ namespace HREngine.Bots
                 temp.AddRange(this.posmoves);
                 havedonesomething = false;
                 Playfield bestold = null;
-                int bestoldval = -20000000;
+                float bestoldval = -20000000;
                 foreach (Playfield p in temp)
                 {
 
@@ -215,12 +215,12 @@ namespace HREngine.Bots
             // Helpfunctions.Instance.logg("find best ");
             if (posmoves.Count >= 1)
             {
-                int bestval = int.MinValue;
+                float bestval = int.MinValue;
                 int bestanzactions = 1000;
                 Playfield bestplay = posmoves[0];//temp[0]
                 foreach (Playfield p in posmoves)//temp
                 {
-                    int val = botBase.getPlayfieldValue(p);
+                    float val = botBase.getPlayfieldValue(p);
                     if (bestval <= val)
                     {
                         if (bestval == val && bestanzactions < p.playactions.Count) continue;
